move SemiRandomStream definition to evoting.go
